Update Isolate package.
No longer contains FunctionRef class since implementations can now send
functions.

Enable error/done listening on the isolate now that the VM implements it.

Add a new abstraction to ports.dart, the "SingleResponseChannel", which is like
the other features in that it creates a Future and a SendPort and forwards the
first value from the SendPort to the Future. This one exposes the two values as
an object with two getters instead of returning one and calling a callback with
the other.

Clean up some formatting suggested by formatter (but far from all).

R=sgjesse@google.com

Review URL: https://codereview.chromium.org//1012053002